configure: Allow using --enable-win64 on aarch64
This renames the installed loader/preloader, to allow them to
be used together with an arm build of Wine.

This is not strictly necessary on its own; Debian packaging
end up with a working wow64 setup on their on by the way they
rename the installed executables anyway, but this allows just
doing a build with --enable-win64 and get it set up the right
way.